prometheus/AKS: Ensure instance names are uniques across all the clusters
By default the instance name is the ip of the node.
The AKS cluster uses the same internal address ranges so a same instance
name can identify a node in different cluster (for example 10.240.0.5)
Using the node name solve the problem as it include a cluster id in
their names.

Related to infra/sysadm-environment#4705