staging/cassandra: try to avoid losing the monitoring when the api server restart
It seems there is a conflict between the Endpoints and the default
EndpointSlice objects. When the api server restart, the server uses the empty
EndpointSlice object and a couple of minutes later, argo sync the cluster
and the monitoring is back.
Directly using an EndpointSlice should help as the EndpointSlice is the
recommended way to this in the recent kubernetes version

Related to swh/infra/sysadm-environment#4874