opentelemetry: Switch from 'limits' configuration key to 'requests'
When using 'limits', exceeding cpu requests is throttled. When pods exceeds memory
usage, the kube scheduler will kill the pod.

Using 'requests' instead will do nothing. It just allows the kube scheduler to determine
whether a pod can be scheduled or not (when pressure exists in the cluster).

We'll be monitoring through the dashboard to better determine the limits to configure
(per cluster).

[1] https://grafana.softwareheritage.org/goto/oPt0RLU4z?orgId=1

Refs. swh/infra/sysadm-environment#4524