Use dynamic index with new opentelemetry-collection version
After testing this within minikube (branch wip-minikube), this adds the required parts
so the oltp collector agents can push logs in daily index (depending on the data
records).

This first update the chart to a version supporting the dynamic index. Then enables the
dynamic index configuration. It configures a suffix and a prefix so the index to store
data is computed out of those.

Internally, the attribute field record 'elasticsearch.index.prefix' is configured with
the environment value (e.g. staging, production, ...). The attribute field
'elasticsearch.index.suffix' is computed out of date time stored in the record field
'asctime'. It's using the day of the year with format '%Y-%m-%d'.

swh/infra/sysadm-environment#4955