Test scrapping of rabbitmq exporter from the cluster
Currently, it's our pergamon node (static infra) which does it. But, as we do not have
an alertmanager, we can't trigger alerts out of the rabbitmq metrics (among other
things). In deploying a service monitor which does the scrape, our metrics can become
accessible to trigger those alerts. In fine, being able to design new alerts to be
notified when stuck listing task issue occur.

Refs. swh/infra/sysadm-environment#5213